Mksquashfs: Use long long in read_bytes()

+++ squashfs-tools/action.h
@@ -318,7 +318,7 @@ extern int eval_empty_actions(struct dir_info *, struct dir_ent *dir_ent);
 extern void eval_move_actions(struct dir_info *, struct dir_ent *);
 extern int eval_prune_actions(struct dir_info *, struct dir_ent *);
 extern void do_move_actions();
-extern int read_bytes(int, void *, int);
+extern long long read_bytes(int, void *, long long);
 extern int actions();
 extern int move_actions();
 extern int empty_actions();
