+++ squashfs-tools/pseudo.h
@@ -83,7 +83,7 @@ struct pseudo {
 	struct pseudo_entry	*name;
 };
 
-extern int read_bytes(int, void *, int);
+extern long long read_bytes(int, void *, long long);
 extern int read_pseudo_definition(char *, char *);
 extern int read_pseudo_file(char *, char *);
 extern struct pseudo *pseudo_subdir(char *, struct pseudo *);
